chapters/appendix-I: Bump license list to v2.6
Generated with:

  $ ./bin/pull-license-list.py

Then I removed the trailing link labels now that the tables are using
inline links.  The inline links make the table a bit wider, but they
are easier to automatically manage because each table is
self-contained (vs. accumulating a set of link lables between all
tables).

The date formatting for "6 Jan 2016" is unfortunate.  I'd prefer
"2016-01-06" or "January 6th, 2016", but the script is just pulling
the raw releaseDate from the JSON, so we get whatever the upstream
data is using.