meeting.ics: Document general, outreach, legal, and technical meetings
The current wiki wording for these meetings makes it sound like
they're pinned to UTC [1], when in fact they are intended to float
with US Pacific Time (America/Los_Angeles) [2].  Having a
machine-redable format makes it possible for everyone's calendar
software to handle those time conversions for them (and allows
non-US-DST participants to avoid internalizing the US DST rules).

I'm not sure the outreach group still exists.  The wiki page doesn't
say anything about the group disolving [3], but the minutes end on
2014-04-24 [4].

The distinction in UberConference channels seems intentional, with
Kate's channel allowing more participants [5].

Future bumps to meeting.ics should bump LAST-MODIFIED [6] or DTSTAMP
[7] for any altered components.  We can't use LAST-MODIFIED in the
VEVENT because VEVENTs require DTSTAMP [8].

The timezone entry is based on the America/New_York example from [9].

The duplicated DESCRIPTION and LOCATION URLs are less DRY, but makes
life easier for folks pulling up the calendar a minute before the
meeting, since they don't have to click through to the wiki pages.
The LOCATION entry [10] seems like a reasonable location for the
conference-call page, so I've used that.  Note that the URI there is a
text value; LOCATION supports URIs in ALTREP [10], but the
UberConference page isn't an "alternate text representation for the
property value" [11].  The URI in DESCRIPTION is a fallback for
clients that don't support/display LOCATION (I'm not sure if any
exist, but it's easy to play it safe).  I couldn't find a more
structured location for the phone number, unless we also wanted to put
that in LOCATION.

com.android.calendar version 7.1.1 only displays the DESCRIPTION body,
and does not provide a link to ALTREP (in fact, I don't see any
instances of ALTREP in the source [12]).  Including the wiki link in
the DESCRIPTION body gives folks using that calendar application an
easy way to get to the wiki pages.

The ICS was validated with [13].

The CRLF line endings are intentional [14], and the .gitattributes
entry ensures we keep them.  The committed files will still have LF
endings, which can confuse 'git diff ...', but you can use
--ignore-space-at-eol to see what really changed.

[1]: https://lists.spdx.org/pipermail/spdx-legal/2017-September/002239.html
     Subject: Meeting times and daylight savings (was: reminder: call
       Thursday)
     Date: Thu, 28 Sep 2017 12:00:08 -0700
     Message-ID: <20170928190008.GA12857@valgrind.tremily.us>
[2]: https://lists.spdx.org/pipermail/spdx-legal/2017-September/002240.html
     Subject: Meeting times and daylight savings (was: reminder: call
       Thursday)
     Date: Thu, 28 Sep 2017 23:29:41 -0600
     Message-ID: <04CF72C4-0256-464C-92F2-8F477535CA5B@jilayne.com>
[3]: https://wiki.spdx.org/view/Outreach_Team
[4]: https://wiki.spdx.org/view/Business_Team/Minutes
[5]: https://lists.spdx.org/pipermail/spdx-legal/2017-September/002236.html
     Subject: CALL NOW: switch uberconference line
     Date: Thu, 28 Sep 2017 11:05:22 -0600
     Message-ID: <7BBCD59E-9139-438A-A2B5-E5BAF1177FEF@jilayne.com>
[6]: https://tools.ietf.org/html/rfc5545#section-3.8.7.3
[7]: https://tools.ietf.org/html/rfc5545#section-3.8.7.2
[8]: https://tools.ietf.org/html/rfc5545#section-3.6.1
[9]: https://tools.ietf.org/html/rfc5545#page-69
[10]: https://tools.ietf.org/html/rfc5545#section-3.8.1.7
[11]: https://tools.ietf.org/html/rfc5545#section-3.2.1
[12]: https://android.googlesource.com/platform/packages/apps/Calendar/+/android-7.1.1_r38
[13]: https://icalendar.org/validator.html
[14]: https://tools.ietf.org/html/rfc5545#section-3.1