meeting.ics: Outreach team meets every other Thursday
In ed7a3369 (meeting.ics: Document general, outreach, legal, and
technical meetings, 2017-09-29), I'd claimed it was the first Thursday
in each month.  I think that was a copy-paste error though, because
since 2017-01-11 the Wiki has had:

> The Outreach Team meets every other Thursday ... starting on 14
> January 2016.

And while ed7a3369 expressed uncertainty about the existence of the
outreach team, they're still going [2].  They're just not posting
minutes to the wiki for some reason.

[1]: https://wiki.spdx.org/index.php?title=Outreach_Team&oldid=4081
     Revised 2017-01-11 14:12 by JackM
[2]: https://github.com/spdx/spdx-spec/pull/42#issuecomment-344682240