meeting.ics: Document the 2018-03-05 general meeting
On Wed, Nov 22, 2017 at 02:46:57PM +0000, Phil Odence wrote [1]:
> As you may know, the Linux Foundation Leadership Summit is in
> Sonoma, March 6-8. Additionally, there will be group meetings on the
> Monday before and Friday after for SPDX and Open Chain respectively.

This has a one-off, physical location, so I've made it a separate
event.  I'm guessing on the time and duration, but we can update those
as we get more information.  I'm also guessing that remote
participants will be able to connect via the usual UberConference
channels; there's no official word on that front yet either.

The physical address is from [2].

The ICS was validated with [3].

[1]: https://lists.spdx.org/pipermail/spdx/2017-November/001152.html
     Subject: SPDX at Leadership Summit in March
     Date: Wed, 22 Nov 2017 14:46:57 +0000
     Message-ID: <C595456F-8430-4DCA-BA9D-64DCD86AD49E@blackducksoftware.com>
[2]: http://events.linuxfoundation.org/events/open-source-leadership-summit
[3]: https://icalendar.org/validator.html