chapters/8: Allow rdf:about for referencing external SPDX IDs
rdf:about seems to be documented in [1].  There are also a lot of
examples using it in [2], but I couldn't find clear a specification
for rdf:about there.  What I'm looking for is a way to reference an
external subject, the annotation tag is the predicate, and the
contents of the annotation element are the object.  That seems to be
what RDFa's 'about' is for [3,4].  [5] is the clearest spec for
attribute semantics that I've found, but I'm not clear how closely
RDFa core 1.1 [3,4,5] matches RDF/XML 1.1 [2].

Also:

* Fix an "annotationg" -> "annotating" typo.
* Remove a doubled annotation element.
* Fix curly quotes in the example I touch.

[1]: https://www.w3.org/TR/rdfa-primer/#alternative-for-setting-the-context-about
[2]: https://www.w3.org/TR/rdf-syntax-grammar/
[3]: https://www.w3.org/TR/rdfa-core/#using-about
[4]: https://www.w3.org/TR/rdfa-core/#A-about
[5]: https://www.w3.org/TR/rdfa-core/#h-s_syntax