chapters/3: Simplify the verification-code algorithm
Instead of using a newline-delimited file to hold and sort hashes,
just store the hashes in an array.  This avoids the need to inject
newlines, sort, and then remove the newlines.  Most languages will
have a native array structure, so there's no need to use the newline
approach as a custom array structure.

Also, use the more intuitive (to me, anyway):

  if file is in "excludes" {
    continue  /* exclude SPDX analysis file(s), etc. */
  }
  operate on the file that's not excluded

instead of the old:

  if file is an 'excludes' file, skip it
    operate on the file that's not excluded
  }