.travis.yml: Transition to a 'deploy' block
This is the approach recommended by Travis [1] for publishing to
GitHub Pages.  Instead of using secret variables to decrypt
deploy-key.enc, the secret variable *is* the deploy token.  That gives
us better security, because Travis will refuse to display the deploy
token in its build logs.

An additional benefit to this commit is that the docs will be built
for all pull requests, but only pushed to gh-pages for master commits.
That will help catch build errors in pull-requests before they are
merged.

Travis docs for configuring $GITHUB_TOKEN are in [2,3,4].  We'll need
to setup that token before merging this commit.

[1]: https://docs.travis-ci.com/user/deployment/pages/
[2]: https://docs.travis-ci.com/user/deployment/pages/#Setting-the-GitHub-token
[3]: https://help.github.com/articles/creating-a-personal-access-token-for-the-command-line/#creating-a-token
[4]: https://docs.travis-ci.com/user/environment-variables#Defining-Variables-in-Repository-Settings