Add examples in non-official formats
As discussed on 2019-04-23 tech team call, this commit adds sample
SPDX files in JSON, XML and YAML format, along with a corresponding
tag-value format version. These are intended for further discussion
and iterating before finalizing any of the new formats.

The versions of the files used in this commit are taken from the
comment at https://github.com/spdx/spdx-spec/issues/55#issuecomment-413063219,
with the only changes being (1) converting CRLF to LF, and (2)
changing the version number in the filenames to 2.0, since these
are in fact SPDX v2.0 files.

Signed-off-by: Steve Winslow <swinslow@gmail.com>