chapters/appendix-VI: Add container registries
A container image can be considered as a package. These packages
are consumed from container registries. Dockerhub is a public
registry that many single use container images are downloaded.
Google Container Registries are used to download container images
for Kubernetes. This change adds a category type called
CONTAINER-IMAGE and adds to it two types: 'Dockerhub' and
'Google Container Registry'.

The most popular local client used to download these images is
Docker. The docker client defaults to Dockerhub. As a result, the
repo doesn't have a 'domain' name associated with it. If anyone
were to receive a tarball of this container image, they will just
find the repo and tag name.

For other registries though, docker will add the domain name to
the container image. For example, if you used docker to work with
images from GCR, you will find the domain name as part of the repo
name.

Other public container registries can be added to this category

Signed-off-by: Nisha K <nishak@vmware.com>