dataLicense can be only one license (CC0-1.0)
According to the description: "By using the SPDX specification ..., you hereby agree that any copyright rights ... shall be subject to the terms of the Creative Commons CC0 1.0 Universal license."

The value of Document Information / dataLicense thus cannot be any arbitrary SPDXExpression, it must be only the expression "CC0-1.0" as shown in the examples.  For consistency with the Description, the schema should specify that any other value of dataLicense is invalid.

(Note: the name of the defined constant CC0-LICENSEID is arbitrary as long as it refers to a string matching the regex ^CC0-1.0$).