Use field name spdxID
All but one of the Document Metadata fields have a consistent name format (spdxVersion, dataLicense, etc).   SPDXID is the name of a type, not a field.  The field name should be spdxID.

The type could be SPDXID rather than string, but since the string can have only a single value, it should be a constant string type matching the regex ^SPDXRef-Document$.   (The name of the constant string type SPDXREF-DOC is arbitrary.)