Fixed a number of small typos
One (additional) reading later, I found and fixed some more typos.

I won't get into British vs. US spelling variations, but I mark here with bold the ones we use (consistently) in the text:

- **acknowledgement(s)** vs. acknowledgment(s)
- organisation(s) vs. **organization(s)**

In appendices I and II, I'd replace _omitable_ with _omissible_, but it's the License List and the Matching Guidelines, so no point in changing it here.