Update the File Checksum field to include data corruption intent
Motivation:

Checksums provide an important tool in combating data corruption;
however, this is currently not reflected in the SPDX description of this
field.

Modification:

The Checksum 'Intent' is updated to include the current goal (avoiding
confusion when multiple versions of a file are present) and the ability
to detect data corruption.

The Description field is updated to provide a more neutral
(intent-independent) description of a checksum.

Result:

SPDX checksum field now supports a more general use of checksums.

Closes: #623
Signed-off-by: Paul Millar <paul.millar@desy.de>