Add in ReleaseDate, BuiltDate, and ValidUntilDate as optional fields
These fields will be in the Usage Profile in SPDX 3.0, however there is a need for them today to support some of the SBOM use cases.    In particular the FDA guidance is asking for an "End of Support" date,  which the ValidUntilDate will satisfy.    Moving these into a 2.3 specific pull request.

Signed-off-by: Kate Stewart <kate.stewart@att.net>