Validate only specific doc examples with SHACL in CI.
The check assumes that JSON example snippets are all parts of a single
JSON document that needs validating, but not all files are written
that way.  To avoid errors that have nothing to do with the
correctness of the examples, we should only check files which meet the
check's assumptions.

Fixes #912.

Signed-off-by: Jeff Licquia <jeff@licquia.org>
Signed-off-by: Arthit Suriyawongkul <arthit@gmail.com>