doc: annex: Validate documentation examples
Updates the security information Annex to conform to the SPDX 3.0 schema
and model. In addition, the CI workflow is updated to use a standalone
script to do the validation (making it easier to validate locally). When
validation documentation files, the script will now detect if the
documentation is a complete document or just fragements of a document.
In the latter case, a wrapper is made around the JSON from the
documentation to make it a valid document for validation (e.g.
"@context", "@graph", and a CreationInfo).

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>
Signed-off-by: Arthit Suriyawongkul <arthit@gmail.com>